Testing output size

+++ ssgan.py
@@ -283,8 +283,8 @@ for epoch in range(num_epochs):
     # For each batch in the dataloader
     for i, data in enumerate(diabetes_loader_train):
         diabetes_data, diabetes_labels, label_mask = data
-        diabetes_data = _to_var(diabetes_data).float()
-        diabetes_labels = _to_var(diabetes_labels).float().squeeze()
+        diabetes_data = _to_var(diabetes_data)
+        diabetes_labels = _to_var(diabetes_labels).long().squeeze()
         label_mask = _to_var(label_mask).float().squeeze()
 
         ##########################
